+++ Tests/SwiftLintFrameworkTests/AutomaticRuleTests.generated.swift
@@ -282,6 +282,12 @@ class ImplicitGetterRuleTests: XCTestCase {
     }
 }
 
+class InclusiveLanguageRuleTests: XCTestCase {
+    func testWithDefaultConfiguration() {
+        verifyRule(InclusiveLanguageRule.description)
+    }
+}
+
 class InertDeferRuleTests: XCTestCase {
     func testWithDefaultConfiguration() {
         verifyRule(InertDeferRule.description)
@@ -567,12 +573,6 @@ class QuickDiscouragedFocusedTestRuleTests: XCTestCase {
 class QuickDiscouragedPendingTestRuleTests: XCTestCase {
     func testWithDefaultConfiguration() {
         verifyRule(QuickDiscouragedPendingTestRule.description)
-    }
-}
-
-class RacistDeclarationRuleTests: XCTestCase {
-    func testWithDefaultConfiguration() {
-        verifyRule(RacistDeclarationRule.description)
     }
 }